python-django-rest-framework/requirements.txt
Django
djangorestframework
markdown
django-filter
drf-nested-routers
django-extensions
ptpython